feat: Working cancel buttons and add messagebox for empty string input

+++ 045_mesa_tuala_F1db/frmDelete.cs
@@ -22,6 +22,14 @@ namespace _045_mesa_tuala_F1db
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
+
+            if (string.IsNullOrWhiteSpace(txtModeldescDelete.Text))
+            {
+                MessageBox.Show("Please input the required field", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return;
+            }
+
+
             int rows_affectd = this.db.delete_record(txtModeldescDelete.Text);
 
             if (rows_affectd > 0)
@@ -33,6 +41,11 @@ namespace _045_mesa_tuala_F1db
             {
                 MessageBox.Show("Failed to add a delete", "Failed", MessageBoxButtons.OK, MessageBoxIcon.Error);
             }
+        }
+
+        private void btnCancel_Click(object sender, EventArgs e)
+        {
+            txtModeldescDelete.Clear();
         }
     }
 }
